+++ server/src/main/java/org/opensearch/identity/AccessToken.java
@@ -1,0 +1,27 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+package org.opensearch.identity;
+
+/**
+ * Tamperproof encapsulation the identity of a subject
+ *
+ * @opensearch.experimental
+ */
+public class AccessToken implements AuthenticationToken {
+
+    /** Actual implementation would be something like:
+
+    public String accessToken;
+    public DateTime accessTokenExpiration;
+    public String refreshToken;
+    public DateTime refreshTokenExpiration;
+
+      */
+
+}